Move sitemaps to last step
