+++ apps/robots-generator/src/components/CommonSettings/CommonSettings.tsx
@@ -1,21 +1,11 @@
-import {
-  Box,
-  Stack,
-  Select,
-  InputLabel,
-  MenuItem,
-  SelectChangeEvent,
-  TextareaAutosize,
-} from "@mui/material";
+import { Box, Stack, InputLabel, TextareaAutosize } from "@mui/material";
 import Input from "@/robots-generator/components/Input";
-import { startOfToday } from "date-fns";
 import InfoIcon from "@mui/icons-material/Info";
 import IconButton from "@mui/material/IconButton";
 import Tooltip from "@mui/material/Tooltip";
 
 import { ChangeEvent, useState } from "react";
 import Timepicker from "@/robots-generator/components/Timepicker";
-import TextArea from "@/robots-generator/components/TextArea";
 import StepperNav from "@/robots-generator/components/StepperNav";
 import { useGlobalState } from "@/robots-generator/context/GlobalContext";
 
@@ -36,7 +26,6 @@ export default function CommonSettings({
   const [cachedDelay, setCachedDelay] = useState(state.cachedDelay);
   const [visitTimeFrom, setVisitTime] = useState(state.visitTimeFrom);
   const [visitTimeTo, setVisitTimeTo] = useState(state.visitTimeTo);
-  const [sitemaps, setSitemaps] = useState(state.sitemaps);
   const [disallowedPaths, setDisallowedPaths] = useState(state.disallowedPaths);
   const [allowedPaths, setAllowedPaths] = useState(state.allowedPaths);
 
@@ -62,11 +51,6 @@ export default function CommonSettings({
     setVisitTimeTo(value);
   };
 
-  const handleSitemapChange = (value: ChangeEvent<HTMLTextAreaElement>) => {
-    const data = value.target.value;
-    setSitemaps(data.split("\n"));
-  };
-
   const handleDisallowedPathsChange = (
     value: ChangeEvent<HTMLTextAreaElement>,
   ) => {
@@ -87,7 +71,6 @@ export default function CommonSettings({
       cachedDelay,
       visitTimeFrom,
       visitTimeTo,
-      sitemaps,
       disallowedPaths,
       allowedPaths,
     });
@@ -264,47 +247,6 @@ export default function CommonSettings({
               placeholder="Enter allowed paths, each path on a new line. They should be relative to the root of your site and end with a /."
               onChange={handleAllowedPathsChange}
               value={allowedPaths.join("\n")}
-              minRows={5}
-              style={{
-                maxHeight: "300px",
-                height: "300px",
-                width: "96%",
-                border: "1px solid #C4C4C4",
-                borderRadius: "5px",
-                padding: "2%",
-                resize: "none",
-                whiteSpace: "pre",
-                lineHeight: "1.5rem",
-                overflowWrap: "normal",
-                overflow: "scroll !important",
-              }}
-            />
-          </Stack>
-          {/* Sitemaps */}
-          <Stack
-            spacing={2}
-            direction="row"
-            alignItems="center"
-            sx={{ width: "100%" }}
-          >
-            <InputLabel
-              sx={{
-                color: "text.primary",
-                fontSize: "16px",
-                width: "100%",
-              }}
-            >
-              Sitemaps
-              <Tooltip title="The sitemap directive specifies the location of your sitemap.">
-                <IconButton>
-                  <InfoIcon />
-                </IconButton>
-              </Tooltip>
-            </InputLabel>
-            <TextareaAutosize
-              placeholder="Enter sitemap URLs, each URL on a new line"
-              onChange={handleSitemapChange}
-              value={sitemaps.join("\n")}
               minRows={5}
               style={{
                 maxHeight: "300px",
