+++ apps/robots-generator/src/components/Sitemaps/index.ts
@@ -1,0 +1,3 @@
+import Sitemaps from "./Sitemaps";
+
+export default Sitemaps;
